+++ relay-event-schema/src/protocol/span/convert.rs
@@ -311,6 +311,7 @@ mod tests {
                 messaging_message_retry_count: ~,
                 messaging_message_receive_latency: ~,
                 messaging_message_body_size: ~,
+                messaging_message_id: ~,
                 other: {},
             },
             sentry_tags: ~,